Unify 03system and 00battle tests
